Memoize the parsing of query string
Using Context.Query was parsing the entire query string each time it was
used. The parsing of the query string should be done once and the Values
stored instead, so that multiple calls to c.Query don't produce a full
map each time. It's possible that this could affect middleware
rewriting, but middleware should really be operating on the http.Request
object and not the Gin context in that case.